refactor: renamed start sync file
