+++ src/network/mod.rs
@@ -2,6 +2,6 @@ pub mod forward;
 pub mod message;
 pub mod peer_ipc;
 pub mod peers_operations;
+pub mod request_filesystem;
 pub mod server;
-pub mod start_sync;
 pub mod watchdogs;
